+++ packages/better_player/lib/src/controls/better_player_material_loading_widget.dart
@@ -11,10 +11,7 @@ class BetterPlayerMaterialLoadingWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     if (controlsConfiguration.loadingWidget != null) {
-      return ColoredBox(
-        color: controlsConfiguration.controlBarColor,
-        child: controlsConfiguration.loadingWidget,
-      );
+      return controlsConfiguration.loadingWidget!;
     }
 
     return CircularProgressIndicator(