+++ packages/better_player/lib/src/controls/better_player_material_progress_bar.dart
@@ -255,7 +255,7 @@ class _ProgressBarPainter extends CustomPainter {
 
   @override
   void paint(Canvas canvas, Size size) {
-    const height = 2.0;
+    const height = 3.0;
 
     canvas.drawRRect(
       RRect.fromRectAndRadius(